Implementing review page, started protecting routes
