+++ dcbr-web/src/store/registration/mutations.ts
@@ -194,4 +194,20 @@ export const mutations: MutationTree<RegistrationState> = {
     state.error = false;
     state.animalIdentification!.otherPermIdType = payload;
   },
+
+  // Terms and Conditions
+  hasAgreed(state: RegistrationState, payload: boolean) {
+    state.error = false;
+    state.termsAndConditions!.hasAgreed = payload;
+  },
+
+  // Route Protection
+  registerFormOk(state: RegistrationState, payload: boolean) {
+    state.error = false;
+    state.routeProtection!.registerFormOk = payload;
+  },
+  reviewFormOk(state: RegistrationState, payload: boolean) {
+    state.error = false;
+    state.routeProtection!.reviewFormOk = payload;
+  }
 };
