+++ dcbr-web/src/store/store.ts
@@ -1,7 +1,7 @@
 import Vue from "vue";
 import Vuex, { StoreOptions } from "vuex";
 
-import { profile, operationDetails, operationLocations, breedingDetails, animalIdentification } from "./registration";
+import { profile, operationDetails, operationLocations, breedingDetails, animalIdentification, termsAndConditions, routeProtection } from "./registration";
 import { RootState } from "./types";
 
 Vue.use(Vuex);
@@ -15,7 +15,9 @@ const store: StoreOptions<RootState> = {
     operationDetails,
     operationLocations,
     breedingDetails,
-    animalIdentification
+    animalIdentification,
+    termsAndConditions,
+    routeProtection
   }
 };
