+++ dcbr-web/src/views/Review.vue
@@ -1,121 +1,23 @@
 <template>
-  <v-app class="grey lighten-4">
-    <v-container>
-      <v-card flat>
-        <v-snackbar v-model="snackbar" absolute top right color="success">
-          <span>Registration successful!</span>
-          <v-icon dark>check_circle</v-icon>
-        </v-snackbar>
-        <v-form ref="form" @submit.prevent="submit">
-          <v-container grid-list-xl fluid>
-            <v-layout wrap>
-              <v-flex xs12>
-                <v-checkbox v-model="form.terms" color="green">
-                  <template v-slot:label>
-                    <div @click.stop>
-                      Do you accept the
-                      <a href="javascript:;" @click.stop="terms = true">terms</a>
-                      and
-                      <a
-                        href="javascript:;"
-                        @click.stop="conditions = true"
-                      >conditions?</a>
-                    </div>
-                  </template>
-                </v-checkbox>
-              </v-flex>
-            </v-layout>
-          </v-container>
-          <v-card-actions>
-            <v-btn flat @click="resetForm">Cancel</v-btn>
-            <v-spacer></v-spacer>
-            <v-btn :disabled="!formIsValid" flat color="primary" type="submit">Register</v-btn>
-          </v-card-actions>
-        </v-form>
-        <v-dialog v-model="terms" width="70%">
-          <v-card>
-            <v-card-title class="title">Terms</v-card-title>
-            <v-card-text v-for="n in 5" :key="n">{{ content }}</v-card-text>
-            <v-card-actions>
-              <v-spacer></v-spacer>
-              <v-btn flat color="purple" @click="terms = false">Ok</v-btn>
-            </v-card-actions>
-          </v-card>
-        </v-dialog>
-        <v-dialog v-model="conditions" width="70%">
-          <v-card>
-            <v-card-title class="title">Conditions</v-card-title>
-            <v-card-text v-for="n in 5" :key="n">{{ content }}</v-card-text>
-            <v-card-actions>
-              <v-spacer></v-spacer>
-              <v-btn flat color="purple" @click="conditions = false">Ok</v-btn>
-            </v-card-actions>
-          </v-card>
-        </v-dialog>
-      </v-card>
-      <!-- snippet-start: CTA To next page -->
-      <v-layout>
-        <v-flex ma-5>
-          <v-btn large round block dark class="primary">
-            <button @click="$router.push('confirmation')">Confirm and register</button>
-          </v-btn>
-        </v-flex>
-      </v-layout>
-      <!-- end: -->
-    </v-container>
-  </v-app>
+  <div class="grey lighten-4">
+    <v-stepper flat>
+      <v-stepper-header>
+        <v-stepper-step complete step="1">Registration</v-stepper-step>
+        <v-divider></v-divider>
+        <v-stepper-step complete editable step="2">Review & Submit</v-stepper-step>
+        <v-divider></v-divider>
+        <v-stepper-step step="3">Confirmation</v-stepper-step>
+      </v-stepper-header>
+    </v-stepper>
+    <RegistrationForm formType="review"/>
+  </div>
 </template>
 
-
-
-
-
-
 <script>
+import RegistrationForm from "@/components/RegistrationForm.vue"
 export default {
-  data() {
-    const defaultForm = Object.freeze({
-      first: "",
-      last: "",
-      bio: "",
-      favoriteAnimal: "",
-      age: null,
-      terms: false
-    });
-    return {
-      form: Object.assign({}, defaultForm),
-      rules: {
-        age: [val => val < 10 || `I don't believe you!`],
-        animal: [val => (val || "").length > 0 || "This field is required"],
-        name: [val => (val || "").length > 0 || "This field is required"]
-      },
-      animals: ["Dog", "Cat", "Rabbit", "Turtle", "Snake"],
-      conditions: false,
-      content: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Integer nec odio. Praesent libero. Sed cursus ante dapibus diam. Sed nisi. Nulla quis sem at nibh elementum imperdiet. Duis sagittis ipsum. Praesent mauris. Fusce nec tellus sed augue semper porta. Mauris massa. Vestibulum lacinia arcu eget nulla. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos. Curabitur sodales ligula in libero. Sed dignissim lacinia nunc.`,
-      snackbar: false,
-      terms: false,
-      defaultForm
-    };
-  },
-  computed: {
-    formIsValid() {
-      return (
-        this.form.first &&
-        this.form.last &&
-        this.form.favoriteAnimal &&
-        this.form.terms
-      );
-    }
-  },
-  methods: {
-    resetForm() {
-      this.form = Object.assign({}, this.defaultForm);
-      this.$refs.form.reset();
-    },
-    submit() {
-      this.snackbar = true;
-      this.resetForm();
-    }
-  }
+  components: {
+    RegistrationForm
+  },    
 };
 </script>
